pyproject.toml
[project]
name = "mill_presenter"
version = "0.1.0"
description = "Grinding mill video analysis and presentation tool"
authors = [
    {name = "MillPresenter Team", email = "team@example.com"},
]
dependencies = [
    "PyQt6>=6.6.0",
    "opencv-python>=4.9.0",
    "av>=11.0.0",
    "numpy>=1.26.0",
    "PyYAML>=6.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Proprietary"}

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
